Attempt at clock module, but it makes things slower.
