+++ firmware/src/lib.rs
@@ -6,3 +6,5 @@ pub mod serial;
 pub mod spi_ws2811;
 pub mod spsc;
 pub mod types;
+pub mod clock;
+
